+++ src/Modes/Visual.ts
@@ -1,3 +1,7 @@
+import {window} from 'vscode';
+import {StaticReflect} from '../LanguageExtensions/StaticReflect';
+import {SymbolMetadata} from '../Symbols/Metadata';
+import {RangeOffset} from '../Types/RangeOffset';
 import {Mode, ModeID} from './Mode';
 import {CommandMap} from '../Mappers/Command';
 import {ActionMoveCursor} from '../Actions/MoveCursor';
@@ -73,13 +77,16 @@ export class ModeVisual extends Mode {
         { keys: 'p', actions: [ActionReplace.selectionsWithRegister], args: {shouldYank: true} },
         { keys: 'P', actions: [ActionReplace.selectionsWithRegister], args: {shouldYank: true} },
 
-        { keys: 'r {char}', actions: [ActionReplace.selectionsWithCharacter] },
+        { keys: 'r {char}', actions: [
+            ActionReplace.selectionsWithCharacter,
+            ActionSelection.shrinkToStarts,
+        ] },
         { keys: '~', actions: [ActionCase.switchSelections] },
 
         { keys: '=', actions: [ActionFilter.Format.bySelections] },
 
-        { keys: '<', actions: [ActionIndent.decrease] },
-        { keys: '>', actions: [ActionIndent.increase] },
+        { keys: '<', actions: [ActionIndent.decrease], args: { isVisualMode: true } },
+        { keys: '>', actions: [ActionIndent.increase], args: { isVisualMode: true } },
 
         { keys: '/', actions: [ActionFind.focusFindWidget] },
 
@@ -117,6 +124,30 @@ export class ModeVisual extends Mode {
         super.enter();
 
         ActionSelection.expandToOne();
+    }
+
+    private _recordedCommandMaps: CommandMap[];
+    get recordedCommandMaps() { return this._recordedCommandMaps; }
+
+    protected onWillCommandMapMakesChanges(map: CommandMap): void {
+        const actions = map.actions.filter(action => {
+            return StaticReflect.getMetadata(SymbolMetadata.Action.shouldSkipOnRepeat, action) !== true;
+        });
+
+        const args = Object.assign({
+            preferedRelativeRange: window.activeTextEditor
+                ? new RangeOffset(window.activeTextEditor.selection)
+                : undefined,
+        }, map.args);
+
+        this._recordedCommandMaps = [
+            {
+                keys: map.keys,
+                actions: actions,
+                args: args,
+                isRepeating: true,
+            }
+        ];
     }
 
 }